Split drawing.c into draw_environment.c and draw_entities.c
