+++ cub3d.h
@@ -6,7 +6,7 @@
 /*   By: iberegsz <iberegsz@student.42vienna.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/21 22:09:04 by andrejarama       #+#    #+#             */
-/*   Updated: 2024/09/01 17:25:02 by iberegsz         ###   ########.fr       */
+/*   Updated: 2024/09/02 11:42:09 by iberegsz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -163,9 +163,18 @@ typedef struct s_vars
 /* Draw_line.c */
 void	draw_line(t_vars *vars, unsigned long color);
 
-/* Drawing.c */
+/* Draw_entities.c */
+void	draw_square(t_vars *vars, int x, int y, int color);
 void	rotate_around_point(t_vars *vars, int *x, int *y);
+void	draw_player(t_vars *vars, unsigned int color);
+void	draw_ray(t_vars *vars, double angle_offset);
+void	draw_ray_segment(t_vars *vars);
+
+/* Draw_environment.c */
+void	draw_floor(t_vars *vars);
+void	draw_ceiling(t_vars *vars);
 void	draw_map(t_vars *vars);
+void	draw_minimap(t_vars *vars);
 
 /* Free_memory_utils.c */
 void	free_memory(char **arr);
